parent_id

+++ database/migrations/2023_08_22_092620_create_categories_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->string('title');
-            $table->unsignedBigInteger('parent_id');
+            $table->unsignedInteger('parent_id');
             $table->timestamps();
         });
     }
